requirements.txt
opencv-python
numpy
scipy
Flask==1.1.2
Pillow==2.2.1
python-dotenv==0.18.0